Updates to build system to better support automake-1.5 - automake-1.4 sets default values for INCLUDES which we can't   overwrite. - automake-1.5 renames this to DEFAULT_INCLUDES and leaves INCLUDES   open for the developer to use.
Thus for automake-1.4 we are forced to 'append' to INCLUDES and in
automake-1.5 we can just set the value to whatever we like.
Unfortunately, the behaviors of the two versions are mutually
incompatible.

The solution I am committing now works for both versions but
automake-1.5 generates a lot of spurious warning messages that are
annoying, but not fatal.